requirements.txt
streamlit>=1.28.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
spotipy>=2.23.0
mutagen>=1.47.0
numpy>=1.24.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1


